Initial commit of builder for Python 3.7
Pretty bare bones at the moment, you must have a requirements.txt in
order for this to work. It's an easy fix to change that, just need to
see if the file exists in the builder script, however I want to get a
copy into source control in case my hd crashes.